Split models.py into multiple files
This commit breaks models.py into multiple files, as this file is
getting larger and larger as time goes by.

We also update job_board/forms.py so that Country is displayed higher
up on the page.
