+++ job_board/context_processors.py
@@ -1,4 +1,5 @@
 from django.contrib.sites.shortcuts import get_current_site
 
+
 def get_site(request):
     return {'current_site': get_current_site(request)}
