+++ job_board/models/country.py
@@ -1,0 +1,16 @@
+from __future__ import unicode_literals
+
+from django.db import models
+from django.utils.encoding import python_2_unicode_compatible
+
+
+@python_2_unicode_compatible
+class Country(models.Model):
+    name = models.CharField(max_length=50, unique=True)
+
+    class Meta:
+        verbose_name_plural = "countries"
+        ordering = ['name']
+
+    def __str__(self):
+        return self.name
