+++ job_board/models/job.py
@@ -9,81 +9,9 @@ from django.template.loader import render_to_string
 from django.utils import timezone
 from django.utils.encoding import python_2_unicode_compatible
 
-
-# NOTE: This uses signals to auto-create a SiteConfig object when a site when a
-#       site is added.  This saves the admin from having to manually create the
-#       site's SiteConfig after a site is added.
-@receiver(models.signals.post_save, sender=Site)
-def generate_site_config(sender, **kwargs):
-    if kwargs.get('created', True):
-        site = kwargs.get('instance')
-        SiteConfig.objects.get_or_create(
-            site=site,
-            admin_email='admin@%s' % site.domain
-        )
-
-
-@python_2_unicode_compatible
-class Category(models.Model):
-    name = models.CharField(max_length=30)
-    site = models.ForeignKey(Site, on_delete=models.CASCADE)
-
-    class Meta:
-        verbose_name_plural = "categories"
-        unique_together = ("name", "site")
-        ordering = ['name']
-
-    def __str__(self):
-        return self.name
-
-    def active_jobs(self):
-        return self.job_set.filter(paid_at__isnull=False) \
-                           .filter(expired_at__isnull=True)
-
-
-@python_2_unicode_compatible
-class Country(models.Model):
-    name = models.CharField(max_length=50, unique=True)
-
-    class Meta:
-        verbose_name_plural = "countries"
-        ordering = ['name']
-
-    def __str__(self):
-        return self.name
-
-
-@python_2_unicode_compatible
-class Company(models.Model):
-    name = models.CharField(max_length=50)
-    url = models.URLField(verbose_name="URL")
-    twitter = models.CharField(
-                  max_length=20,
-                  help_text="Please leave empty if none"
-              )
-    country = models.ForeignKey(
-                  Country,
-                  blank=True,
-                  null=True,
-                  help_text="Please leave empty if 100% virtual"
-              )
-    site = models.ForeignKey(Site, on_delete=models.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-    class Meta:
-        verbose_name_plural = "companies"
-        unique_together = ("name", "site")
-        ordering = ['name']
-
-    def __str__(self):
-        return self.name
-
-    def active_jobs(self):
-        return self.job_set.filter(paid_at__isnull=False,
-                                   expired_at__isnull=True)
-
-    def paid_jobs(self):
-        return self.job_set.filter(paid_at__isnull=False)
+from job_board.models.category import Category
+from job_board.models.company import Company
+from job_board.models.country import Country
 
 
 @python_2_unicode_compatible
@@ -158,15 +86,3 @@ class Job(models.Model):
 
     def __str__(self):
         return self.title
-
-
-@python_2_unicode_compatible
-class SiteConfig(models.Model):
-    expire_after = models.SmallIntegerField(default=30)
-    # NOTE: We set a default here, but we will override this with a more
-    #       suitable default when we create the SiteConfig instance
-    admin_email = models.EmailField(default='admin@site')
-    site = models.ForeignKey(Site, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.site.name
